Rename scenarios/aws/vm/ec2VM to scenarios/aws/vm/ec2vm
