+++ scenarios/aws/vm/ec2vm/ec2VM.go
